@@ -3,7 +3,8 @@ package consts
 import "time"
 
 const (
-	ServerAddr = ":3000"
+	ServerAddr              = ":3000"
+	ServerReadHeaderTimeout = 60 * time.Second
 
 	RequestLimit = 100
 	WindowLength = 1 * time.Minute
@@ -18,4 +19,8 @@ const (
 	TmpLog = "./tmp/logs/server.log"
 
 	TimeoutShutdown = 10 * time.Second
+
+	VkOauthRedirectUrl     = "/api/auth/vk/callback"
+	OkOauthRedirectUrl     = "/api/auth/ok/callback"
+	YandexOauthRedirectUrl = "/api/auth/yandex/callback"
 )
