@@ -3,8 +3,6 @@ package market
 type HandlerMarket struct {
 }
 
-func NewHandler(
-) HandlerMarket {
-	return HandlerMarket{
-	}
+func NewHandler() HandlerMarket {
+	return HandlerMarket{}
 }
